webserver/requirements-taihe-test.txt
gevent
git+https://github.com/esnme/ultrajson.git
git+https://github.com/esnme/ultramysql.git
git+https://github.com/gf0842wf/mprpc.git
bottle
gevent-websocket
websocket-client
bottle-websocket
pymongo